adding summary and tasks tabs; refactoring code; fixing css
