+++ webapp/src/components/rhs/incident_details.scss
@@ -3,7 +3,7 @@ $font-family: Open Sans;
 .IncidentDetails {
   display: flex;
   flex-direction: column;
-  padding: 2rem 2rem 10rem;
+  padding: 2rem 2rem 12rem;
 
   .side-by-side {
     display: flex;
